初回実行時ディレクトリが無いのでエラー終了していた

+++ publish.sh
@@ -2,6 +2,8 @@
 
 set -eux
 
-rm -r output
+if [ -d output ]; then
+  rm -r output
+fi
 jbake -b
 rsync -av -e ssh ./output/ himeji-cs@himeji-cs.jp:www/blog2
